+++ airbyte-cdk/python/unit_tests/sources/declarative/retrievers/test_simple_retriever.py
@@ -3,7 +3,7 @@
 #
 
 from typing import Mapping
-from unittest.mock import MagicMock, patch
+from unittest.mock import MagicMock, Mock, patch
 
 import airbyte_cdk.sources.declarative.requesters.error_handlers.response_status as response_status
 import pytest
@@ -11,7 +11,7 @@ import requests
 from airbyte_cdk.models import AirbyteLogMessage, AirbyteMessage, Level, SyncMode, Type
 from airbyte_cdk.sources.declarative.auth.declarative_authenticator import NoAuth
 from airbyte_cdk.sources.declarative.exceptions import ReadException
-from airbyte_cdk.sources.declarative.incremental import DatetimeBasedCursor
+from airbyte_cdk.sources.declarative.incremental import Cursor, DatetimeBasedCursor
 from airbyte_cdk.sources.declarative.partition_routers import SinglePartitionRouter
 from airbyte_cdk.sources.declarative.requesters.error_handlers.response_action import ResponseAction
 from airbyte_cdk.sources.declarative.requesters.error_handlers.response_status import ResponseStatus
@@ -24,6 +24,10 @@ from airbyte_cdk.sources.declarative.retrievers.simple_retriever import (
     _response_to_airbyte_message,
 )
 from airbyte_cdk.sources.streams.http.http import HttpStream
+
+A_SLICE_STATE = {"slice_state": "slice state value"}
+A_STREAM_SLICE = {"stream slice": "slice value"}
+A_STREAM_STATE = {"stream state": "state value"}
 
 primary_key = "pk"
 records = [{"id": 1}, {"id": 2}]
@@ -48,14 +52,14 @@ def test_simple_retriever_full(mock_http_stream):
     record_selector = MagicMock()
     record_selector.select_records.return_value = records
 
-    stream_slicer = MagicMock()
+    cursor = MagicMock(spec=Cursor)
     stream_slices = [{"date": "2022-01-01"}, {"date": "2022-01-02"}]
-    stream_slicer.stream_slices.return_value = stream_slices
+    cursor.stream_slices.return_value = stream_slices
 
     response = requests.Response()
 
     underlying_state = {"date": "2021-01-01"}
-    stream_slicer.get_stream_state.return_value = underlying_state
+    cursor.get_stream_state.return_value = underlying_state
 
     requester.get_authenticator.return_value = NoAuth({})
     url_base = "https://airbyte.io"
@@ -87,7 +91,8 @@ def test_simple_retriever_full(mock_http_stream):
         requester=requester,
         paginator=paginator,
         record_selector=record_selector,
-        stream_slicer=stream_slicer,
+        stream_slicer=cursor,
+        cursor=cursor,
         parameters={},
         config={},
     )
@@ -98,13 +103,13 @@ def test_simple_retriever_full(mock_http_stream):
     assert retriever.state == underlying_state
     assert retriever.next_page_token(response) == next_page_token
     assert retriever.request_params(None, None, None) == request_params
-    assert retriever.stream_slices(sync_mode=SyncMode.incremental) == stream_slices
+    assert retriever.stream_slices() == stream_slices
 
     assert retriever._last_response is None
-    assert retriever._last_records is None
+    assert retriever._records_from_last_response is None
     assert retriever.parse_response(response, stream_state={}) == records
     assert retriever._last_response == response
-    assert retriever._last_records == records
+    assert retriever._records_from_last_response == records
 
     assert retriever.http_method == "GET"
     assert not retriever.raise_on_http_errors
@@ -185,10 +190,10 @@ def test_simple_retriever_with_request_response_log_last_records(mock_http_strea
     )
 
     assert retriever._last_response is None
-    assert retriever._last_records is None
+    assert retriever._records_from_last_response is None
     assert retriever.parse_response(response, stream_state={}) == request_response_logs
     assert retriever._last_response == response
-    assert retriever._last_records == request_response_logs
+    assert retriever._records_from_last_response == request_response_logs
 
     [r for r in retriever.read_records(SyncMode.full_refresh)]
     paginator.reset.assert_called()
@@ -569,7 +574,7 @@ def test_path(test_name, requester_path, paginator_path, expected_path):
                 type=Type.LOG,
                 log=AirbyteLogMessage(
                     level=Level.INFO,
-                    message='request:{"url": "https://airbyte.io/", "http_method": "GET", "headers": {"Content-Type": "application/json", "Content-Length": "24"}, "body": {"b1": "v1", "b2": "v2"}}',
+                    message='request:{"url": "https://airbyte.io/", "http_method": "GET", "headers": {"Content-Type": "application/json", "Content-Length": "24"}, "body": "{\\"b1\\": \\"v1\\", \\"b2\\": \\"v2\\"}"}',
                 ),
             ),
         ),
@@ -585,7 +590,7 @@ def test_path(test_name, requester_path, paginator_path, expected_path):
                 type=Type.LOG,
                 log=AirbyteLogMessage(
                     level=Level.INFO,
-                    message='request:{"url": "https://airbyte.io/?p1=v1&p2=v2", "http_method": "GET", "headers": {"Content-Type": "application/json", "h1": "v1", "Content-Length": "24"}, "body": {"b1": "v1", "b2": "v2"}}',
+                    message='request:{"url": "https://airbyte.io/?p1=v1&p2=v2", "http_method": "GET", "headers": {"Content-Type": "application/json", "h1": "v1", "Content-Length": "24"}, "body": "{\\"b1\\": \\"v1\\", \\"b2\\": \\"v2\\"}"}',
                 ),
             ),
         ),
@@ -593,7 +598,7 @@ def test_path(test_name, requester_path, paginator_path, expected_path):
             "test_get_request_with_request_body_data",
             HttpMethod.GET,
             "https://airbyte.io",
-            {"Content-Type": "application/json"},
+            {"Content-Type": "application/x-www-form-urlencoded"},
             {},
             {},
             {"b1": "v1", "b2": "v2"},
@@ -601,7 +606,7 @@ def test_path(test_name, requester_path, paginator_path, expected_path):
                 type=Type.LOG,
                 log=AirbyteLogMessage(
                     level=Level.INFO,
-                    message='request:{"url": "https://airbyte.io/", "http_method": "GET", "headers": {"Content-Type": "application/json", "Content-Length": "11"}, "body": {"b1": "v1", "b2": "v2"}}',
+                    message='request:{"url": "https://airbyte.io/", "http_method": "GET", "headers": {"Content-Type": "application/x-www-form-urlencoded", "Content-Length": "11"}, "body": "b1=v1&b2=v2"}',
                 ),
             ),
         ),
@@ -715,45 +720,79 @@ def test_limit_stream_slices():
         config={},
     )
 
-    truncated_slices = list(retriever.stream_slices(sync_mode=SyncMode.incremental, stream_state=None))
+    truncated_slices = list(retriever.stream_slices())
 
     assert truncated_slices == _generate_slices(maximum_number_of_slices)
 
 
 @pytest.mark.parametrize(
-    "test_name, last_records, records, expected_stream_slicer_update_count",
+    "test_name, first_greater_than_second",
     [
-        ("test_two_records", [{"id": -1}], records, 2),
-        ("test_no_records", [{"id": -1}], [], 1),
-        ("test_no_records_no_previous_records", [], [], 0)
-    ]
+        ("test_first_greater_than_second", True),
+        ("test_second_greater_than_first", False),
+    ],
 )
-def test_read_records_updates_stream_slicer_once_if_no_records(test_name, last_records, records, expected_stream_slicer_update_count):
-    with patch.object(HttpStream, "_read_pages", return_value=iter(records)):
-        requester = MagicMock()
-        paginator = MagicMock()
-        record_selector = MagicMock()
-        stream_slicer = MagicMock()
+def test_when_read_records_then_cursor_close_slice_with_greater_record(test_name, first_greater_than_second):
+    requester = MagicMock()
+    paginator = MagicMock()
+    record_selector = MagicMock()
+    first_record = Mock()
+    second_record = Mock()
+    records = [first_record, second_record]
+    record_selector.select_records.return_value = records
+    cursor = MagicMock(spec=Cursor)
+    cursor.is_greater_than_or_equal.return_value = first_greater_than_second
 
-        retriever = SimpleRetriever(
-            name="stream_name",
-            primary_key=primary_key,
-            requester=requester,
-            paginator=paginator,
-            record_selector=record_selector,
-            stream_slicer=stream_slicer,
-            parameters={},
-            config={},
-        )
-        retriever._last_records = last_records
+    retriever = SimpleRetriever(
+        name="stream_name",
+        primary_key=primary_key,
+        requester=requester,
+        paginator=paginator,
+        record_selector=record_selector,
+        stream_slicer=cursor,
+        cursor=cursor,
+        parameters={},
+        config={},
+    )
+    stream_slice = {"repository": "airbyte"}
 
-        list(retriever.read_records(sync_mode=SyncMode.incremental, stream_slice={"repository": "airbyte"}))
+    with patch.object(HttpStream, "_read_pages", return_value=iter([first_record, second_record]), side_effect=lambda _, __, ___: retriever.parse_records(request=MagicMock(), response=MagicMock(), stream_state=None, stream_slice=stream_slice)):
+        list(retriever.read_records(sync_mode=SyncMode.incremental, stream_slice=stream_slice))
+        cursor.close_slice.assert_called_once_with(stream_slice, first_record if first_greater_than_second else second_record)
 
-        assert stream_slicer.update_cursor.call_count == expected_stream_slicer_update_count
+
+def parse_two_pages_and_return_records(retriever, stream_slice, records):
+    list(retriever.parse_records(request=MagicMock(), response=MagicMock(), stream_state=None, stream_slice=stream_slice))
+    list(retriever.parse_records(request=MagicMock(), response=MagicMock(), stream_state=None, stream_slice=stream_slice))
+    return records
 
 
 def _generate_slices(number_of_slices):
     return [{"date": f"2022-01-0{day + 1}"} for day in range(number_of_slices)]
+
+
+@patch.object(HttpStream, "_read_pages", return_value=iter([]))
+def test_given_state_selector_when_read_records_use_slice_state(http_stream_read_pages):
+    requester = MagicMock()
+    paginator = MagicMock()
+    record_selector = MagicMock()
+    cursor = MagicMock(spec=Cursor)
+    cursor.select_state = MagicMock(return_value=A_SLICE_STATE)
+
+    retriever = SimpleRetriever(
+        name="stream_name",
+        primary_key=primary_key,
+        requester=requester,
+        paginator=paginator,
+        record_selector=record_selector,
+        stream_slicer=cursor,
+        cursor=cursor,
+        parameters={},
+        config={},
+    )
+    list(retriever.read_records(SyncMode.incremental, stream_slice=A_STREAM_SLICE))
+
+    http_stream_read_pages.assert_called_once_with(retriever.parse_records, A_STREAM_SLICE, A_SLICE_STATE)
 
 
 def test_emit_log_request_response_messages():