+++ tools/ci_connector_ops/ci_connector_ops/pipelines/commands/groups/metadata.py
@@ -1,7 +1,6 @@
 #
 # Copyright (c) 2023 Airbyte, Inc., all rights reserved.
 #
-import logging
 
 import anyio
 import click
@@ -16,14 +15,9 @@ from ci_connector_ops.pipelines.pipelines.metadata import (
 from ci_connector_ops.pipelines.utils import (
     DaggerPipelineCommand,
     get_all_metadata_files,
-    get_modified_metadata_files,
     get_expected_metadata_files,
+    get_modified_metadata_files,
 )
-from rich.logging import RichHandler
-
-logging.basicConfig(level=logging.INFO, format="%(name)s: %(message)s", datefmt="[%X]", handlers=[RichHandler(rich_tracebacks=True)])
-logger = logging.getLogger(__name__)
-
 
 # MAIN GROUP
 
@@ -88,6 +82,7 @@ def upload(ctx: click.Context, gcs_bucket_name: str, modified_only: bool) -> boo
         ctx.obj["git_branch"],
         ctx.obj["git_revision"],
         ctx.obj.get("gha_workflow_run_url"),
+        ctx.obj.get("dagger_logs_url"),
         ctx.obj.get("pipeline_start_timestamp"),
         ctx.obj.get("ci_context"),
         metadata_to_upload,
@@ -113,6 +108,7 @@ def deploy_orchestrator(ctx: click.Context) -> bool:
         ctx.obj["git_branch"],
         ctx.obj["git_revision"],
         ctx.obj.get("gha_workflow_run_url"),
+        ctx.obj.get("dagger_logs_url"),
         ctx.obj.get("pipeline_start_timestamp"),
         ctx.obj.get("ci_context"),
     )
@@ -136,6 +132,7 @@ def test_lib(ctx: click.Context) -> bool:
         ctx.obj["git_branch"],
         ctx.obj["git_revision"],
         ctx.obj.get("gha_workflow_run_url"),
+        ctx.obj.get("dagger_logs_url"),
         ctx.obj.get("pipeline_start_timestamp"),
         ctx.obj.get("ci_context"),
     )
@@ -150,6 +147,7 @@ def test_orchestrator(ctx: click.Context) -> bool:
         ctx.obj["git_branch"],
         ctx.obj["git_revision"],
         ctx.obj.get("gha_workflow_run_url"),
+        ctx.obj.get("dagger_logs_url"),
         ctx.obj.get("pipeline_start_timestamp"),
         ctx.obj.get("ci_context"),
     )